+++ CHANGELOG.md
@@ -13,6 +13,15 @@ and this plugin adheres loosely to [Semantic Versioning](https://semver.org/).
 
 TBD
 
+## [1.2] - 2026-02-26
+
+Clear cache in admin bar enhanced
+
+* JavaScript logic rewritten to not rely on jQuery.
+* CSS created with loading.gif embedded directly into css for performance.
+* Errors and success message simplified for display in admin bar menu.
+* Icon source: [Loading icon.gif](https://commons.wikimedia.org/wiki/File:YouTube_loading_symbol_3_%28transparent%29.gif)
+
 ## [1.1] - 2026-01-30
 
 Clear cache functionality added.